Rename lib/translations/translations.js -> lib/translations/all.js
