+++ lib/translations/all.js
