+++ lib/translations/index.js
@@ -4,7 +4,7 @@ var config = require('lib/config');
 var log = require('debug')('democracyos:translations');
 var available = require('./available');
 var missing = require('./missing');
-var translations = require('./translations');
+var translations = require('./all');
 var baseLocale = require('./base-locale');
 
 // Throw a Warning if there are missing translations
