+++ lib/translations/missing.js
@@ -1,4 +1,4 @@
-var translations = require('./translations');
+var translations = require('./all');
 var available = require('./available');
 var baseLocale = require('./base-locale');
